perf(webp): i16 coefficient pipeline (stage B) — block arrays + kernel i16 in/out
The encoder's DCT-coefficient carriers and the fused kernel family now run
on i16 end-to-end, matching libwebp's int16_t pipeline (#38 speed work):

- LumaBlockResult::coeffs, luma_blocks/u_blocks/v_blocks, y1_quant/
  y1_dequant/uv_quant/uv_dequant, I4BlockResult::dequantized,
  best_dequantized: [i32;N] -> [i16;N] (DCT output, levels, and
  dequantized level*q are all i16-bounded by construction).
- ftransform_from_u8_4x4* return [i16;16] (the SSE2 kernel computed i16
  internally and sign-extended at return — that tail is deleted).
- quantize_dequantize_block_* / _ac_only / luma16_ac_arcane: i16 in/out
  on every tier (SSE2/NEON/WASM were already 16-bit inside; the input
  packs and output sign-extend unpacks are deleted — 12 fewer
  instructions per SSE2 call).
- idct_add_residue_i16 (+ _sse2_inner) and idct4x4_i16: i16 variants for
  the encoder; the i32 originals remain for the decoder. Shared #[rite]
  core (idct_add_res_dc/full_sse2) so both load styles reuse one body.
- add_residue_i16 replaces the encoder's i32 add_residue (deleted along
  with its dispatch tiers; test retargeted).
- sse4x4_with_residual residual param -> &[i16;16] on all tiers (drops
  the 4-load+2-pack prelude).
- diffuse_chroma_dc_inplace operates on i16 blocks.
- Trellis stays i32 in/out this slice: callers widen at the (m5/m6-only,
  out-of-line) boundaries — i4_trellis_quantize_recon, the I16 final
  pass, i16_rd_trellis_quantize, and the recorder's verify/requant arms.
- Dead code deleted: the four unused legacy record paths in residuals.rs
  (encode_residual_data / encode_coefficients / record_residual_stats /
  record_residual_tokens, ~690 lines), Plane enum, DCT token constants +
  DCT_TOKEN_TREE, DCT_CAT_BASE.

Byte gate: output_hash COMBINED 958f376a6c8b118f — identical before/
after (21 images x both cost models x m0-6 x q{5,50,75,90} + alpha/
sharp/lossless). Suites: x86-64 741 passed, i686 715 passed, wasm32
builds, clippy x3 clean.

+++ src/lib.rs
@@ -331,12 +331,12 @@ pub mod __test_helpers {
     }
 
     /// Forward DCT used by the encoder; dispatched (SSE2 / NEON / etc).
-    pub fn ftransform_from_u8_4x4_dispatch(src: &[u8; 16], ref_: &[u8; 16]) -> [i32; 16] {
+    pub fn ftransform_from_u8_4x4_dispatch(src: &[u8; 16], ref_: &[u8; 16]) -> [i16; 16] {
         crate::common::transform::ftransform_from_u8_4x4(src, ref_)
     }
 
     /// Scalar reference forward DCT.
-    pub fn ftransform_from_u8_4x4_scalar(src: &[u8; 16], ref_: &[u8; 16]) -> [i32; 16] {
+    pub fn ftransform_from_u8_4x4_scalar(src: &[u8; 16], ref_: &[u8; 16]) -> [i16; 16] {
         crate::common::transform::ftransform_from_u8_4x4_scalar(src, ref_)
     }
 
@@ -371,11 +371,11 @@ pub mod __test_helpers {
 
     /// Dispatched fused quantize+dequantize used by the encoder.
     pub fn quantize_dequantize_block_dispatch(
-        coeffs: &[i32; 16],
+        coeffs: &[i16; 16],
         matrix: &crate::encoder::quantize::VP8Matrix,
         use_sharpen: bool,
-        quantized: &mut [i32; 16],
-        dequantized: &mut [i32; 16],
+        quantized: &mut [i16; 16],
+        dequantized: &mut [i16; 16],
     ) -> bool {
         crate::encoder::quantize::quantize_dequantize_block_simd(
             coeffs,
@@ -389,10 +389,10 @@ pub mod __test_helpers {
     /// Scalar reference fused quantize+dequantize (does NOT use sharpen,
     /// matching the scalar dispatch tier's behavior).
     pub fn quantize_dequantize_block_scalar(
-        coeffs: &[i32; 16],
+        coeffs: &[i16; 16],
         matrix: &crate::encoder::quantize::VP8Matrix,
-        quantized: &mut [i32; 16],
-        dequantized: &mut [i32; 16],
+        quantized: &mut [i16; 16],
+        dequantized: &mut [i16; 16],
     ) -> bool {
         crate::encoder::quantize::quantize_dequantize_block_scalar(
             coeffs,
@@ -471,7 +471,7 @@ pub mod __test_helpers {
     pub fn sse4x4_with_residual_dispatch(
         src: &[u8; 16],
         pred: &[u8; 16],
-        dequantized: &[i32; 16],
+        dequantized: &[i16; 16],
     ) -> u32 {
         crate::encoder::vp8::mode_selection::__test_only_sse4x4_with_residual_dispatch(
             src,
@@ -484,11 +484,11 @@ pub mod __test_helpers {
     pub fn sse4x4_with_residual_scalar(
         src: &[u8; 16],
         pred: &[u8; 16],
-        dequantized: &[i32; 16],
+        dequantized: &[i16; 16],
     ) -> u32 {
         let mut sum = 0u32;
         for i in 0..16 {
-            let rec = (i32::from(pred[i]) + dequantized[i]).clamp(0, 255) as u8;
+            let rec = (i32::from(pred[i]) + i32::from(dequantized[i])).clamp(0, 255) as u8;
             let d = (src[i] as i32 - rec as i32).unsigned_abs();
             sum += d * d;
         }
